+++ src/models/metrics.rs
@@ -2,5 +2,5 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Metrics {
-    pub gen_utime: u64,
+    pub gen_utime: u32,
 }
